- tokens page moved to folder /tokens - token-table component created to support multiple tables (for pegged & engine tokens) - relabeling & moving functionality to token-table component - commenting out "load more" as currently we fetch all tokens
